+++ src-ui/app/models/vrchat.ts
@@ -1,5 +1,18 @@
+import { Avatar } from 'vrchat/dist';
+
 export interface WorldContext {
   playerCount: number;
   instanceId?: string;
   loaded: boolean;
+}
+
+export type AvatarEx = Avatar & {
+  favoriteGroup?: string;
+  favoriteId?: string;
+};
+
+export interface PersistedAvatar {
+  id: string;
+  imageUrl: string;
+  name: string;
 }